refactor(agent): inline reflection support-file collection
